@@ -1,24 +1,25 @@
 name: cupertino_http
-description: >
+version: 1.4.0
+description: >-
   A macOS/iOS Flutter plugin that provides access to the Foundation URL
   Loading System.
-version: 1.1.0-wip
 repository: https://github.com/dart-lang/http/tree/master/pkgs/cupertino_http
 
 environment:
-  sdk: ^3.0.0
-  flutter: '>=3.10.0'  # If changed, update test matrix.
+  sdk: ^3.3.0
+  flutter: '>=3.19.0'  # If changed, update test matrix.
 
 dependencies:
-  ffi: ^2.0.1
+  async: ^2.5.0
+  ffi: ^2.1.0
   flutter:
     sdk: flutter
-  http: '>=0.13.4 <2.0.0'
-  meta: ^1.7.0
+  http: ^1.2.0
+  web_socket: ^0.1.0
 
 dev_dependencies:
-  dart_flutter_team_lints: ^1.0.0
-  ffigen: ^8.0.2
+  dart_flutter_team_lints: ^2.0.0
+  ffigen: ^9.0.1
 
 flutter:
   plugin:
