@@ -1,10 +1,10 @@
 name: cancellation_token_http
-version: 2.0.0
+version: 2.1.0
 homepage: https://github.com/JonathanPeterCole/dart-cancellation-token-http/tree/main/pkgs/http
 description: A fork of Dart's HTTP package with request cancellation using cancellation_token.
 
 environment:
-  sdk: ^3.0.0
+  sdk: ^3.3.0
 
 platforms:
   android:
@@ -19,9 +19,10 @@ dependencies:
   cancellation_token: ^2.0.0
   http_parser: ^4.0.0
   meta: ^1.3.0
+  web: ^0.5.0
 
 dev_dependencies:
-  dart_flutter_team_lints: ^1.0.0
+  dart_flutter_team_lints: ^2.0.0
   fake_async: ^1.2.0
   http_client_conformance_tests:
     path: ../http_client_conformance_tests/
