@@ -10,7 +10,7 @@ import 'package:stream_channel/stream_channel.dart';
 import 'package:test/test.dart';
 
 import 'request_body_server_vm.dart'
-    if (dart.library.html) 'request_body_server_web.dart';
+    if (dart.library.js_interop) 'request_body_server_web.dart';
 
 class _Plus2Decoder extends Converter<List<int>, String> {
   @override
@@ -40,16 +40,16 @@ class _Plus2Encoding extends Encoding {
 /// 'POST'.
 void testRequestBody(Client client) {
   group('request body', () {
-    late final String host;
-    late final StreamChannel<Object?> httpServerChannel;
-    late final StreamQueue<Object?> httpServerQueue;
+    late String host;
+    late StreamChannel<Object?> httpServerChannel;
+    late StreamQueue<Object?> httpServerQueue;
 
-    setUpAll(() async {
+    setUp(() async {
       httpServerChannel = await startServer();
       httpServerQueue = StreamQueue(httpServerChannel.stream);
-      host = 'localhost:${await httpServerQueue.next}';
+      host = 'localhost:${await httpServerQueue.nextAsInt}';
     });
-    tearDownAll(() => httpServerChannel.sink.add(null));
+    tearDown(() => httpServerChannel.sink.add(null));
 
     test('client.post() with string body', () async {
       await client.post(Uri.http(host, ''), body: 'Hello World!');
@@ -152,5 +152,172 @@ void testRequestBody(Client client) {
       expect(serverReceivedContentType, ['image/png; charset=plus2']);
       expect(serverReceivedBody.codeUnits, [1, 2, 3, 4, 5]);
     });
+
+    test('client.send() with stream containing empty lists', () async {
+      final request = StreamedRequest('POST', Uri.http(host, ''));
+      request.headers['Content-Type'] = 'image/png';
+      request.sink.add([]);
+      request.sink.add([]);
+      request.sink.add([1]);
+      request.sink.add([2]);
+      request.sink.add([]);
+      request.sink.add([3, 4]);
+      request.sink.add([]);
+      request.sink.add([5]);
+      // ignore: unawaited_futures
+      request.sink.close();
+      await client.send(request);
+
+      final serverReceivedContentType = await httpServerQueue.next;
+      final serverReceivedBody = await httpServerQueue.next as String;
+
+      expect(serverReceivedContentType, ['image/png']);
+      expect(serverReceivedBody.codeUnits, [1, 2, 3, 4, 5]);
+    });
+
+    test('client.send() with slow stream', () async {
+      Stream<List<int>> stream() async* {
+        await Future<void>.delayed(const Duration(milliseconds: 100));
+        yield [1];
+        await Future<void>.delayed(const Duration(milliseconds: 100));
+        yield [2];
+        await Future<void>.delayed(const Duration(milliseconds: 100));
+        yield [3];
+        await Future<void>.delayed(const Duration(milliseconds: 100));
+        yield [4];
+        await Future<void>.delayed(const Duration(milliseconds: 100));
+        yield [5];
+        await Future<void>.delayed(const Duration(milliseconds: 100));
+        yield [6, 7, 8];
+        await Future<void>.delayed(const Duration(milliseconds: 100));
+        yield [9, 10];
+        await Future<void>.delayed(const Duration(milliseconds: 100));
+      }
+
+      final request = StreamedRequest('POST', Uri.http(host, ''));
+      request.headers['Content-Type'] = 'image/png';
+
+      stream().listen(request.sink.add,
+          onError: request.sink.addError, onDone: request.sink.close);
+      await client.send(request);
+
+      final serverReceivedContentType = await httpServerQueue.next;
+      final serverReceivedBody = await httpServerQueue.next as String;
+
+      expect(serverReceivedContentType, ['image/png']);
+      expect(serverReceivedBody.codeUnits, [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]);
+    });
+
+    test('client.send() with stream that raises', () async {
+      Stream<List<int>> stream() async* {
+        yield [0];
+        yield [1];
+        throw ArgumentError('this is a test');
+      }
+
+      final request = StreamedRequest('POST', Uri.http(host, ''));
+      request.headers['Content-Type'] = 'image/png';
+
+      stream().listen(request.sink.add,
+          onError: request.sink.addError, onDone: request.sink.close);
+
+      await expectLater(client.send(request),
+          throwsA(anyOf(isA<ArgumentError>(), isA<ClientException>())));
+    });
+
+    test('client.send() GET with empty stream', () async {
+      final request = StreamedRequest('GET', Uri.http(host, ''));
+      request.headers['Content-Type'] = 'image/png';
+      // ignore: unawaited_futures
+      request.sink.close();
+
+      final response = await client.send(request);
+      expect(response.statusCode, 200);
+
+      final serverReceivedContentType = await httpServerQueue.next;
+      final serverReceivedBody = await httpServerQueue.next as String;
+
+      expect(serverReceivedContentType, ['image/png']);
+      expect(serverReceivedBody.codeUnits, <int>[]);
+    });
+
+    test('client.send() GET with stream containing only empty lists', () async {
+      final request = StreamedRequest('GET', Uri.http(host, ''));
+      request.headers['Content-Type'] = 'image/png';
+      request.sink.add([]);
+      request.sink.add([]);
+      request.sink.add([]);
+      // ignore: unawaited_futures
+      request.sink.close();
+
+      final response = await client.send(request);
+      expect(response.statusCode, 200);
+
+      final serverReceivedContentType = await httpServerQueue.next;
+      final serverReceivedBody = await httpServerQueue.next as String;
+
+      expect(serverReceivedContentType, ['image/png']);
+      expect(serverReceivedBody.codeUnits, <int>[]);
+    });
+
+    test('client.send() with persistentConnection', () async {
+      // Do five requests to verify that the connection persistance logic is
+      // correct.
+      for (var i = 0; i < 5; ++i) {
+        final request = Request('POST', Uri.http(host, ''))
+          ..headers['Content-Type'] = 'text/plain; charset=utf-8'
+          ..persistentConnection = true
+          ..body = 'Hello World $i';
+
+        final response = await client.send(request);
+        expect(response.statusCode, 200);
+
+        final serverReceivedContentType = await httpServerQueue.next;
+        final serverReceivedBody = await httpServerQueue.next as String;
+
+        expect(serverReceivedContentType, ['text/plain; charset=utf-8']);
+        expect(serverReceivedBody, 'Hello World $i');
+      }
+    });
+
+    test('client.send() with persistentConnection and body >64K', () async {
+      // 64KiB is special for the HTTP network API:
+      // https://fetch.spec.whatwg.org/#http-network-or-cache-fetch
+      // See https://github.com/dart-lang/http/issues/977
+      final body = ''.padLeft(64 * 1024, 'XYZ');
+
+      final request = Request('POST', Uri.http(host, ''))
+        ..headers['Content-Type'] = 'text/plain; charset=utf-8'
+        ..persistentConnection = true
+        ..body = body;
+
+      final response = await client.send(request);
+      expect(response.statusCode, 200);
+
+      final serverReceivedContentType = await httpServerQueue.next;
+      final serverReceivedBody = await httpServerQueue.next as String;
+
+      expect(serverReceivedContentType, ['text/plain; charset=utf-8']);
+      expect(serverReceivedBody, body);
+    });
+
+    test('client.send() GET with non-empty stream', () async {
+      final request = StreamedRequest('GET', Uri.http(host, ''));
+      request.headers['Content-Type'] = 'image/png';
+      request.sink.add('Hello World!'.codeUnits);
+      // ignore: unawaited_futures
+      request.sink.close();
+
+      final response = await client.send(request);
+      expect(response.statusCode, 200);
+
+      final serverReceivedContentType = await httpServerQueue.next;
+      final serverReceivedBody = await httpServerQueue.next as String;
+
+      expect(serverReceivedContentType, ['image/png']);
+      expect(serverReceivedBody, 'Hello World!');
+      // using io passes, on web body is not transmitted, on cupertino_http
+      // exception.
+    }, skip: 'unclear semantics for GET requests with body');
   });
 }
