@@ -5,6 +5,7 @@
 import 'dart:io';
 
 import 'package:cronet_http/cronet_http.dart';
+import 'package:http/http.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:test/test.dart';
 
@@ -31,7 +32,7 @@ void testCache() {
     });
 
     test('disabled', () async {
-      final engine = await CronetEngine.build(cacheMode: CacheMode.disabled);
+      final engine = CronetEngine.build(cacheMode: CacheMode.disabled);
       final client = CronetClient.fromCronetEngine(engine);
       await client.get(Uri.parse('http://localhost:${server.port}'));
       await client.get(Uri.parse('http://localhost:${server.port}'));
@@ -39,7 +40,7 @@ void testCache() {
     });
 
     test('memory', () async {
-      final engine = await CronetEngine.build(
+      final engine = CronetEngine.build(
           cacheMode: CacheMode.memory, cacheMaxSize: 1024 * 1024);
       final client = CronetClient.fromCronetEngine(engine);
       await client.get(Uri.parse('http://localhost:${server.port}'));
@@ -48,7 +49,7 @@ void testCache() {
     });
 
     test('disk', () async {
-      final engine = await CronetEngine.build(
+      final engine = CronetEngine.build(
           cacheMode: CacheMode.disk,
           cacheMaxSize: 1024 * 1024,
           storagePath: (await Directory.systemTemp.createTemp()).absolute.path);
@@ -59,7 +60,7 @@ void testCache() {
     });
 
     test('diskNoHttp', () async {
-      final engine = await CronetEngine.build(
+      final engine = CronetEngine.build(
           cacheMode: CacheMode.diskNoHttp,
           cacheMaxSize: 1024 * 1024,
           storagePath: (await Directory.systemTemp.createTemp()).absolute.path);
@@ -76,14 +77,14 @@ void testInvalidConfigurations() {
   group('invalidConfigurations', () {
     test('no storagePath', () async {
       expect(
-          () async => await CronetEngine.build(
+          () async => CronetEngine.build(
               cacheMode: CacheMode.disk, cacheMaxSize: 1024 * 1024),
           throwsArgumentError);
     });
 
     test('non-existing storagePath', () async {
       expect(
-          () async => await CronetEngine.build(
+          () async => CronetEngine.build(
               cacheMode: CacheMode.disk,
               cacheMaxSize: 1024 * 1024,
               storagePath: '/a/b/c/d'),
@@ -112,10 +113,38 @@ void testUserAgent() {
     });
 
     test('userAgent', () async {
-      final engine = await CronetEngine.build(userAgent: 'fake-agent');
+      final engine = CronetEngine.build(userAgent: 'fake-agent');
       await CronetClient.fromCronetEngine(engine)
           .get(Uri.parse('http://localhost:${server.port}'));
       expect(requestHeaders['user-agent'], ['fake-agent']);
+    });
+  });
+}
+
+void testEngineClose() {
+  group('engine close', () {
+    test('multiple close', () {
+      CronetEngine.build()
+        ..close()
+        ..close();
+    });
+
+    test('request after close', () async {
+      final closedEngine = CronetEngine.build()..close();
+      final client = CronetClient.fromCronetEngine(closedEngine);
+      await expectLater(() => client.get(Uri.https('example.com', '/')),
+          throwsA(isA<ClientException>()));
+    });
+
+    test('engine owned close', () {
+      final engine = CronetEngine.build();
+      CronetClient.fromCronetEngine(engine, closeEngine: true).close();
+    });
+
+    test('engine not owned close', () {
+      final engine = CronetEngine.build();
+      CronetClient.fromCronetEngine(engine, closeEngine: false).close();
+      engine.close();
     });
   });
 }
@@ -126,4 +155,5 @@ void main() {
   testCache();
   testInvalidConfigurations();
   testUserAgent();
+  testEngineClose();
 }
