@@ -5,24 +5,31 @@
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:cached_network_image/cached_network_image.dart';
 import 'package:cronet_http/cronet_http.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart';
+import 'package:http/io_client.dart';
+import 'package:http_image_provider/http_image_provider.dart';
+import 'package:provider/provider.dart';
 
 import 'book.dart';
 
 void main() {
-  var clientFactory = Client.new; // Constructs the default client.
+  final Client httpClient;
   if (Platform.isAndroid) {
-    Future<CronetEngine>? engine;
-    clientFactory = () {
-      engine ??= CronetEngine.build(
-          cacheMode: CacheMode.memory, userAgent: 'Book Agent');
-      return CronetClient.fromCronetEngineFuture(engine!);
-    };
+    final engine = CronetEngine.build(
+        cacheMode: CacheMode.memory,
+        cacheMaxSize: 2 * 1024 * 1024,
+        userAgent: 'Book Agent');
+    httpClient = CronetClient.fromCronetEngine(engine, closeEngine: true);
+  } else {
+    httpClient = IOClient(HttpClient()..userAgent = 'Book Agent');
   }
-  runWithClient(() => runApp(const BookSearchApp()), clientFactory);
+
+  runApp(Provider<Client>(
+      create: (_) => httpClient,
+      child: const BookSearchApp(),
+      dispose: (_, client) => client.close()));
 }
 
 class BookSearchApp extends StatelessWidget {
@@ -46,20 +53,23 @@ class HomePage extends StatefulWidget {
 
 class _HomePageState extends State<HomePage> {
   List<Book>? _books;
+  String? _lastQuery;
+  late Client _client;
 
   @override
   void initState() {
     super.initState();
+    _client = context.read<Client>();
   }
 
   // Get the list of books matching `query`.
   // The `get` call will automatically use the `client` configurated in `main`.
   Future<List<Book>> _findMatchingBooks(String query) async {
-    final response = await get(
+    final response = await _client.get(
       Uri.https(
         'www.googleapis.com',
         '/books/v1/volumes',
-        {'q': query, 'maxResults': '40', 'printType': 'books'},
+        {'q': query, 'maxResults': '20', 'printType': 'books'},
       ),
     );
 
@@ -68,6 +78,7 @@ class _HomePageState extends State<HomePage> {
   }
 
   void _runSearch(String query) async {
+    _lastQuery = query;
     if (query.isEmpty) {
       setState(() {
         _books = null;
@@ -76,6 +87,9 @@ class _HomePageState extends State<HomePage> {
     }
 
     final books = await _findMatchingBooks(query);
+    // Avoid the situation where a slow-running query finishes late and
+    // replaces newer search results.
+    if (query != _lastQuery) return;
     setState(() {
       _books = books;
     });
@@ -127,10 +141,10 @@ class _BookListState extends State<BookList> {
         itemBuilder: (context, index) => Card(
           key: ValueKey(widget.books[index].title),
           child: ListTile(
-            leading: CachedNetworkImage(
-                placeholder: (context, url) =>
-                    const CircularProgressIndicator(),
-                imageUrl: widget.books[index].imageUrl),
+            leading: Image(
+                image: HttpImage(
+                    widget.books[index].imageUrl.replace(scheme: 'https'),
+                    client: context.read<Client>())),
             title: Text(widget.books[index].title),
             subtitle: Text(widget.books[index].description),
           ),
