@@ -6,4 +6,4 @@ import 'base_client.dart';
 
 /// Implemented in `browser_client.dart` and `io_client.dart`.
 BaseClient createClient() => throw UnsupportedError(
-    'Cannot create a client without dart:html or dart:io.');
+    'Cannot create a client without dart:js_interop or dart:io.');
