@@ -10,5 +10,9 @@ import 'package:http_client_conformance_tests/http_client_conformance_tests.dart
 import 'package:test/test.dart';
 
 void main() {
-  testAll(IOClient.new);
+  testAll(
+    IOClient.new, preservesMethodCase: false, // https://dartbug.com/54187
+    canReceiveSetCookieHeaders: true,
+    canSendCookieHeaders: true,
+  );
 }
