@@ -47,15 +47,22 @@ void main() {
     final response = await request.send();
 
     expect(response.statusCode, equals(302));
+    expect(
+        response,
+        isA<http.BaseResponseWithUrl>()
+            .having((r) => r.url, 'url', serverUrl.resolve('/redirect')));
   });
 
   test('with redirects', () async {
     final request = http.Request('GET', serverUrl.resolve('/redirect'));
     final response = await request.send();
-
     expect(response.statusCode, equals(200));
     final bytesString = await response.stream.bytesToString();
     expect(bytesString, parse(containsPair('path', '/')));
+    expect(
+        response,
+        isA<http.BaseResponseWithUrl>()
+            .having((r) => r.url, 'url', serverUrl.resolve('/')));
   });
 
   test('exceeding max redirects', () async {
