@@ -10,11 +10,16 @@ import 'dart:io';
 
 import 'package:dart_style/dart_style.dart';
 
-const vm = '''// Generated by generate_server_wrappers.dart. Do not edit.
+const _export = '''export 'server_queue_helpers.dart'
+    show StreamQueueOfNullableObjectExtension;''';
+
+const _vm = '''// Generated by generate_server_wrappers.dart. Do not edit.
 
 import 'package:stream_channel/stream_channel.dart';
 
 import '<server_file_placeholder>';
+
+$_export
 
 /// Starts the redirect test HTTP server in the same process.
 Future<StreamChannel<Object?>> startServer() async {
@@ -24,10 +29,12 @@ Future<StreamChannel<Object?>> startServer() async {
 }
 ''';
 
-const web = '''// Generated by generate_server_wrappers.dart. Do not edit.
+const _web = '''// Generated by generate_server_wrappers.dart. Do not edit.
 
 import 'package:stream_channel/stream_channel.dart';
 import 'package:test/test.dart';
+
+$_export
 
 /// Starts the redirect test HTTP server out-of-process.
 Future<StreamChannel<Object?>> startServer() async => spawnHybridUri(Uri(
@@ -41,11 +48,11 @@ void main() async {
 
   files.where((file) => file.path.endsWith('_server.dart')).forEach((file) {
     final vmPath = file.path.replaceAll('_server.dart', '_server_vm.dart');
-    File(vmPath).writeAsStringSync(formatter.format(vm.replaceAll(
+    File(vmPath).writeAsStringSync(formatter.format(_vm.replaceAll(
         '<server_file_placeholder>', file.uri.pathSegments.last)));
 
     final webPath = file.path.replaceAll('_server.dart', '_server_web.dart');
-    File(webPath).writeAsStringSync(formatter.format(web.replaceAll(
+    File(webPath).writeAsStringSync(formatter.format(_web.replaceAll(
         '<server_file_placeholder>', file.uri.pathSegments.last)));
   });
 }
