@@ -8,7 +8,7 @@ import 'package:stream_channel/stream_channel.dart';
 import 'package:test/test.dart';
 
 import 'compressed_response_body_server_vm.dart'
-    if (dart.library.html) 'compressed_response_body_server_web.dart';
+    if (dart.library.js_interop) 'compressed_response_body_server_web.dart';
 
 /// Tests that the [Client] correctly implements HTTP responses with compressed
 /// bodies.
@@ -32,7 +32,7 @@ void testCompressedResponseBody(Client client) async {
     setUpAll(() async {
       httpServerChannel = await startServer();
       httpServerQueue = StreamQueue(httpServerChannel.stream);
-      host = 'localhost:${await httpServerQueue.next}';
+      host = 'localhost:${await httpServerQueue.nextAsInt}';
     });
     tearDownAll(() => httpServerChannel.sink.add(null));
 
