@@ -8,7 +8,7 @@ import 'package:stream_channel/stream_channel.dart';
 import 'package:test/test.dart';
 
 import 'redirect_server_vm.dart'
-    if (dart.library.html) 'redirect_server_web.dart';
+    if (dart.library.js_interop) 'redirect_server_web.dart';
 
 /// Tests that the [Client] correctly implements HTTP redirect logic.
 ///
@@ -23,9 +23,20 @@ void testRedirect(Client client, {bool redirectAlwaysAllowed = false}) async {
     setUpAll(() async {
       httpServerChannel = await startServer();
       httpServerQueue = StreamQueue(httpServerChannel.stream);
-      host = 'localhost:${await httpServerQueue.next}';
+      host = 'localhost:${await httpServerQueue.nextAsInt}';
     });
     tearDownAll(() => httpServerChannel.sink.add(null));
+
+    test('no redirect', () async {
+      final request = Request('GET', Uri.http(host, '/'))
+        ..followRedirects = false;
+      final response = await client.send(request);
+      expect(response.statusCode, 200);
+      expect(response.isRedirect, false);
+      if (response case BaseResponseWithUrl(url: final url)) {
+        expect(url, Uri.http(host, '/'));
+      }
+    });
 
     test('disallow redirect', () async {
       final request = Request('GET', Uri.http(host, '/1'))
@@ -33,6 +44,21 @@ void testRedirect(Client client, {bool redirectAlwaysAllowed = false}) async {
       final response = await client.send(request);
       expect(response.statusCode, 302);
       expect(response.isRedirect, true);
+      if (response case BaseResponseWithUrl(url: final url)) {
+        expect(url, Uri.http(host, '/1'));
+      }
+    }, skip: redirectAlwaysAllowed ? 'redirects always allowed' : false);
+
+    test('disallow redirect, 0 maxRedirects', () async {
+      final request = Request('GET', Uri.http(host, '/1'))
+        ..followRedirects = false
+        ..maxRedirects = 0;
+      final response = await client.send(request);
+      expect(response.statusCode, 302);
+      expect(response.isRedirect, true);
+      if (response case BaseResponseWithUrl(url: final url)) {
+        expect(url, Uri.http(host, '/1'));
+      }
     }, skip: redirectAlwaysAllowed ? 'redirects always allowed' : false);
 
     test('allow redirect', () async {
@@ -41,9 +67,12 @@ void testRedirect(Client client, {bool redirectAlwaysAllowed = false}) async {
       final response = await client.send(request);
       expect(response.statusCode, 200);
       expect(response.isRedirect, false);
+      if (response case BaseResponseWithUrl(url: final url)) {
+        expect(url, Uri.http(host, '/'));
+      }
     });
 
-    test('allow redirect, 0 maxRedirects, ', () async {
+    test('allow redirect, 0 maxRedirects', () async {
       final request = Request('GET', Uri.http(host, '/1'))
         ..followRedirects = true
         ..maxRedirects = 0;
@@ -60,6 +89,9 @@ void testRedirect(Client client, {bool redirectAlwaysAllowed = false}) async {
       final response = await client.send(request);
       expect(response.statusCode, 200);
       expect(response.isRedirect, false);
+      if (response case BaseResponseWithUrl(url: final url)) {
+        expect(url, Uri.http(host, '/'));
+      }
     }, skip: redirectAlwaysAllowed ? 'redirects always allowed' : false);
 
     test('too many redirects', () async {
