@@ -2,6 +2,5 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library;
-
+export 'src/java_client.dart';
 export 'src/java_http_base.dart';
